FeedBack의 score를 double로 변경

+++ src/main/java/com/example/SchoolLunchReport/statistics/controller/dto/response/StatisticsResponse.java
@@ -2,8 +2,6 @@ package com.example.SchoolLunchReport.statistics.controller.dto.response;
 
 import com.example.SchoolLunchReport.statistics.domain.type.PeriodType;
 import java.util.List;
-import java.util.Objects;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 
 @Builder
@@ -13,7 +11,7 @@ public record StatisticsResponse(
 ) {
 
     public record ScoreCount(
-        int score,
+        double score,
         long count
     ) {
